ShipDetail側でのカスタムHook作成
